# myblog/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/user/register": {
            "post": {
                "description": "用户注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "type": "string",
                        "description": "验证参数Bearer和token空格拼接",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "交款查询参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user_api.RegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.DataStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user_service.RegisterOut"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.DataStruct"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.DataStruct": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "inner_msg": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "ts": {
                    "type": "integer"
                }
            }
        },
        "user_api.RegisterReq": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_num": {
                    "type": "string"
                },
                "rand_code": {
                    "type": "string"
                }
            }
        },
        "user_service.RegisterOut": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "expireTime": {
                    "type": "integer"
                },
                "refreshToken": {
                    "type": "string"
                },
                "status": {
                    "description": "1为正常 2 为该账户已完成注册走 登录流程 -1 通用异常",
                    "type": "integer"
                }
            }
        }
    }
}